requirements.txt
langchain
ipykernel
langserve
streamlit
python-dotenv
langchain-groq
langchain-community
fastapi